Welcome to the New and Improved Main!
This is pretty much the full skeleton of the program so far. Main objectives right now are to get the menu printing and switch on the input so we can navigate the application. Honestly, the menu will probably be pretty different, but we need to move around the different functions and this seems like the easiest way.